+++ scalr/resource_scalr_workspace_migrate.go
@@ -2,6 +2,7 @@ package scalr
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/scalr/go-scalr"
 )
 
 func resourceScalrWorkspaceResourceV0() *schema.Resource {
@@ -329,5 +330,208 @@ func resourceScalrWorkspaceResourceV2() *schema.Resource {
 
 func resourceScalrWorkspaceStateUpgradeV2(rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
 	delete(rawState, "queue_all_runs")
+	return rawState, nil
+}
+
+func resourceScalrWorkspaceResourceV3() *schema.Resource {
+	return &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"name": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+
+			"environment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+
+			"vcs_provider_id": {
+				Type:          schema.TypeString,
+				Optional:      true,
+				ConflictsWith: []string{"module_version_id"},
+				RequiredWith:  []string{"vcs_repo"},
+			},
+			"module_version_id": {
+				Type:          schema.TypeString,
+				Optional:      true,
+				ConflictsWith: []string{"vcs_provider_id", "vcs_repo"},
+			},
+			"agent_pool_id": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+
+			"auto_apply": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  false,
+			},
+
+			"var_files": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
+
+			"operations": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  true,
+			},
+
+			"terraform_version": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+
+			"working_directory": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Default:  "",
+			},
+
+			"hooks": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"pre_init": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Default:  "",
+						},
+
+						"pre_plan": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Default:  "",
+						},
+
+						"post_plan": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Default:  "",
+						},
+
+						"pre_apply": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Default:  "",
+						},
+
+						"post_apply": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Default:  "",
+						},
+					},
+				},
+			},
+
+			"has_resources": {
+				Type:     schema.TypeBool,
+				Computed: true,
+			},
+
+			"vcs_repo": {
+				Type:          schema.TypeList,
+				Optional:      true,
+				MinItems:      1,
+				MaxItems:      1,
+				ConflictsWith: []string{"module_version_id"},
+				RequiredWith:  []string{"vcs_provider_id"},
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"identifier": {
+							Type:     schema.TypeString,
+							Required: true,
+						},
+
+						"branch": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+
+						"path": {
+							Type:       schema.TypeString,
+							Default:    "",
+							Optional:   true,
+							Deprecated: "The attribute `vcs-repo.path` is deprecated. Use working-directory and trigger-prefixes instead.",
+						},
+
+						"trigger_prefixes": {
+							Type:     schema.TypeList,
+							Elem:     &schema.Schema{Type: schema.TypeString},
+							Optional: true,
+							Computed: true,
+						},
+
+						"dry_runs_enabled": {
+							Type:     schema.TypeBool,
+							Optional: true,
+							Default:  true,
+						},
+						"ingress_submodules": {
+							Type:     schema.TypeBool,
+							Optional: true,
+							Default:  false,
+						},
+					},
+				},
+			},
+
+			"created_by": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"username": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"email": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"full_name": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+					},
+				},
+			},
+			"run_operation_timeout": {
+				Type:     schema.TypeInt,
+				Optional: true,
+			},
+			"provider_configuration": {
+				Type:     schema.TypeSet,
+				Optional: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"id": {
+							Type:     schema.TypeString,
+							Required: true,
+						},
+						"alias": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func resourceScalrWorkspaceStateUpgradeV3(rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
+	if rawState["operations"].(bool) {
+		rawState["execution_mode"] = scalr.WorkspaceExecutionModeRemote
+	} else {
+		rawState["execution_mode"] = scalr.WorkspaceExecutionModeLocal
+	}
 	return rawState, nil
 }